+++ test/fonts/test_set_font.py
@@ -107,6 +107,7 @@ def test_set_font_core_font_attributes():
         "ut": 50,
         "cw": fpdf_charwidths["courier"],
         "fontkey": "courier",
+        "emphasis": 0,
     }
     assert pdf.fonts["times"] == {
         "i": 2,
@@ -116,6 +117,7 @@ def test_set_font_core_font_attributes():
         "ut": 50,
         "cw": fpdf_charwidths["times"],
         "fontkey": "times",
+        "emphasis": 0,
     }